+++ src/main/java/xuyang/dev/xuyangapi/config/OpenApiConfig.java
@@ -10,20 +10,16 @@ import java.util.List;
 @Configuration
 public class OpenApiConfig {
 
-//    @Value("${openapi.servers.local}")
-//    private String localUrl;
-//
-//    @Value("${openapi.servers.prod}")
-//    private String prodUrl;
-//
-//    @Bean
-//    public OpenAPI customOpenAPI() {
-//        return new OpenAPI()
-//                .servers(List.of(
-//                        new Server().url(localUrl).description("local"),
-//                        new Server().url(prodUrl).description("production gateway")
-//                ));
-//    }
+    @Value("${openapi.servers.prod}")
+    private String swaggerUrl;
+
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .servers(List.of(
+                        new Server().url(swaggerUrl).description("prod")
+                ));
+    }
 
 }